+++ ouds_core/lib/components/button/internal/ouds_button_padding_modifier.dart
@@ -21,39 +21,44 @@ class OudsButtonPaddingModifier {
   static EdgeInsetsDirectional getPadding(
     BuildContext context,
     OudsButtonLayout layout,
+    OudsNavigationButtonLayout? navigationLayout,
+    OudsButtonComponent componentType,
   ) {
-    final theme = OudsTheme.of(context);
+    final buttonTokens = OudsTheme.of(context).componentsTokens(context).button;
+    final isNavigationButton =
+        componentType == OudsButtonComponent.navigationButton;
+    final isNextLayout =
+        isNavigationButton &&
+        navigationLayout == OudsNavigationButtonLayout.next;
+    final isPreviousLayout =
+        isNavigationButton &&
+        navigationLayout == OudsNavigationButtonLayout.previous;
     switch (layout) {
       case OudsButtonLayout.iconOnly:
         return EdgeInsetsDirectional.all(
-          theme.componentsTokens(context).button.spaceInsetIconOnlyDefault,
+          buttonTokens.spaceInsetIconOnlyDefault,
         );
       case OudsButtonLayout.iconAndText:
         return EdgeInsetsDirectional.only(
-          top: theme.componentsTokens(context).button.spacePaddingBlockDefault,
-          end: theme
-              .componentsTokens(context)
-              .button
-              .spacePaddingInlineEndIconStartDefault,
-          bottom: theme
-              .componentsTokens(context)
-              .button
-              .spacePaddingBlockDefault,
-          start: theme
-              .componentsTokens(context)
-              .button
-              .spacePaddingInlineIconStartDefault,
+          top: buttonTokens.spacePaddingBlockDefault,
+          end:
+              isNextLayout //case navigation button with next layout
+              ? buttonTokens.spacePaddingInlineChevronEndDefault
+              : //case navigation button with previous layout or normal button
+                buttonTokens.spacePaddingInlineEndIconStartDefault,
+          bottom: buttonTokens.spacePaddingBlockDefault,
+          start:
+              isNextLayout // case navigation button with next layout
+              ? buttonTokens.spacePaddingInlineStartIconEndDefault
+              : isPreviousLayout // case navigation button with previous layout
+              ? buttonTokens.spacePaddingInlineChevronStartDefault
+              : //normal button
+                buttonTokens.spacePaddingInlineIconStartDefault,
         );
       case OudsButtonLayout.textOnly:
         return EdgeInsetsDirectional.symmetric(
-          vertical: theme
-              .componentsTokens(context)
-              .button
-              .spacePaddingBlockDefault,
-          horizontal: theme
-              .componentsTokens(context)
-              .button
-              .spacePaddingInlineIconNoneDefault,
+          vertical: buttonTokens.spacePaddingBlockDefault,
+          horizontal: buttonTokens.spacePaddingInlineIconNoneDefault,
         );
     }
   }